blobsync/requirements.txt
azure-storage-blob==12.19.0
azure-identity==1.14.0
python-dotenv==1.0.0
pydantic==2.5.0
